requirements.txt
argcomplete==3.5.3
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
pipx==1.7.1
platformdirs==4.3.6
pycparser==2.22
PyPDF2==3.0.1
pypdfium2==4.30.1
python-docx==1.1.2
typing_extensions==4.12.2
userpath==1.9.2
Werkzeug==3.1.3
